get_famous_people_photos/requirements.txt
# facenet
tensorflow == 1.7.0
requests
numpy
scipy
scikit-learn
opencv_python
facenet_sandberg
h5py
matplotlib
Pillow
psutil
dhash
pybktree
